+++ pkg/repositories/errors.go
@@ -6,6 +6,15 @@ var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
 }
 
+var fileNotFoundError = &microerror.Error{
+	Kind: "fileNotFoundError",
+}
+
+// IsFileNotFoundError returns true if error is fileNotFoundError.
+func IsFileNotFoundError(err error) bool {
+	return microerror.Cause(err) == fileNotFoundError
+}
+
 var repositoryNotFoundError = &microerror.Error{
 	Kind: "repositoryNotFoundError",
 }